Initial commit
The code was previously published in a gist but belongs in a repository so it’s easier to view, update and download the data.

What's New in Version 1.1:

* Cleaner and (mostly) more accurate data.
* Fixes text encoding bugs and issues with garbled fields.
* Fixes issues with different ways peoples names have been typed.
* Improved output format including party name and ISO date values.
* Does a better job of fetching info about guests political party affiliation (pulls directly from Wikipedia and combines structured data with scraping).
* Still fails to handle getting party data for some people with more complicated statuses (e.g. Alex Salmond, who was obviously SNP, but was expelled breifly in 1982, rejoined and then resigned in 2018).

A future update might combine the current code with the old code from https://gist.github.com/iaincollins/cff498c0541ebe7c51ea903238d1d9a0/revisions or an enhanced parser to better deal with special cases.

It may be a good idea to try and clean up party names as well. One way to fix this that doesn’t involve code would be to update the episode guide and member information so that links in Wikipedia are more consistant.